@@ -1,18 +1,16 @@
 name: keicy_avatar_image
 description: A new Flutter package project.
 version: 0.0.1
-author:
 homepage:
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
-  flutter: ">=1.17.0 <2.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
 
-  shimmer: ^1.1.2
+  shimmer:
   
 dev_dependencies:
   flutter_test:
